internal(xychart/components/series): move non-series components into private/ directory
