+++ packages/visx-xychart/src/components/series/BarGroup.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { PositionScale } from '@visx/shape/lib/types';
-import BaseBarGroup, { BaseBarGroupProps } from './BaseBarGroup';
-import Bars from './Bars';
+import BaseBarGroup, { BaseBarGroupProps } from './private/BaseBarGroup';
+import Bars from './private/Bars';
 
 export default function BarGroup<
   XScale extends PositionScale,
