+++ packages/visx-xychart/src/components/series/BarSeries.tsx
@@ -1,7 +1,7 @@
 import { AxisScale } from '@visx/axis';
 import React from 'react';
-import BaseBarSeries, { BaseBarSeriesProps } from './BaseBarSeries';
-import Bars from './Bars';
+import BaseBarSeries, { BaseBarSeriesProps } from './private/BaseBarSeries';
+import Bars from './private/Bars';
 
 function BarSeries<XScale extends AxisScale, YScale extends AxisScale, Datum extends object>({
   ...props
