+++ packages/visx-xychart/src/components/series/BarStack.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { PositionScale } from '@visx/shape/lib/types';
-import BaseBarStack, { BaseBarStackProps } from './BaseBarStack';
-import Bars from './Bars';
+import BaseBarStack, { BaseBarStackProps } from './private/BaseBarStack';
+import Bars from './private/Bars';
 
 export default function BarStack<
   XScale extends PositionScale,
