+++ packages/visx-xychart/src/components/series/private/AnimatedBars.tsx
