+++ packages/visx-xychart/src/components/series/private/Bars.tsx
