+++ packages/visx-xychart/src/components/series/private/BaseBarGroup.tsx
