+++ packages/visx-xychart/src/components/series/private/BaseBarSeries.tsx
